+++ tests/extension_cases/mod.rs
@@ -28,7 +28,7 @@ use datafusion::arrow::array::RecordBatch;
 use datafusion::arrow::util::pretty::pretty_format_batches;
 use datafusion::sql::parser::DFParser;
 use datafusion_common::Result;
-use dft::execution::ExecutionContext;
+use dft::execution::local::ExecutionContext;
 use dft::{config::AppConfig, execution::AppType};
 use futures::{StreamExt, TryStreamExt};
 use log::debug;
